move history to internal
