+++ drivers/mac/driver_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/murlokswarm/app"
-	"github.com/murlokswarm/app/tests"
+	"github.com/murlokswarm/app/internal/tests"
 )
 
 func TestDriver(t *testing.T) {
