+++ drivers/test/driver_test.go
@@ -5,7 +5,7 @@ import (
 	"testing"
 
 	"github.com/murlokswarm/app"
-	"github.com/murlokswarm/app/tests"
+	"github.com/murlokswarm/app/internal/tests"
 )
 
 func TestDriver(t *testing.T) {
