+++ drivers/test/window.go
@@ -19,7 +19,7 @@ type Window struct {
 	id          uuid.UUID
 	factory     app.Factory
 	markup      app.Markup
-	history     app.History
+	history     core.History
 	lastFocus   time.Time
 	component   app.Compo
 	x           float64
@@ -35,8 +35,8 @@ func newWindow(d *Driver, c app.WindowConfig) (app.Window, error) {
 	var markup app.Markup = html.NewMarkup(d.factory)
 	markup = app.ConcurrentMarkup(markup)
 
-	history := app.NewHistory()
-	history = app.ConcurrentHistory(history)
+	history := core.NewHistory()
+	history = core.ConcurrentHistory(history)
 
 	win := &Window{
 		id:          uuid.New(),
