+++ event_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/murlokswarm/app"
-	"github.com/murlokswarm/app/tests"
+	"github.com/murlokswarm/app/internal/tests"
 )
 
 func TestEventRegistry(t *testing.T) {
