+++ html/markup_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/murlokswarm/app"
-	"github.com/murlokswarm/app/tests"
+	"github.com/murlokswarm/app/internal/tests"
 )
 
 func TestMarkup(t *testing.T) {
