+++ internal/tests/resources/logo.png
