+++ internal/tests/window.go
